+++ src/main/java/com/itsweb/backend/member/controller/MemberController.java
@@ -1,5 +1,8 @@
-package com.itsweb.backend.member;
+package com.itsweb.backend.member.controller;
 
+import com.itsweb.backend.member.SessionConst;
+import com.itsweb.backend.member.domain.Member;
+import com.itsweb.backend.member.service.MemberService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
@@ -7,14 +10,24 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api")
-@RequiredArgsConstructor
 @Slf4j
-public class LoginController {
+@RequiredArgsConstructor
+public class MemberController {
+
     private final MemberService memberService;
+    private final PasswordEncoder bCryptPasswordEncoder;
+
+    @PostMapping("/join")
+    public ResponseEntity<?> join(@Valid @RequestBody MemberDTO memberDTO) {
+        Member member = new Member();
+        member.signUp(memberDTO.getUserId(), memberDTO.getUsername(), memberDTO.getPassword(),bCryptPasswordEncoder);
+        return memberService.join(member);
+    }
 
     @PostMapping("/login")
     public ResponseEntity<?> login(@Valid @RequestBody LoginDTO loginDTO, HttpServletRequest request) {
@@ -22,10 +35,9 @@ public class LoginController {
         if (loginMember == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("아이디 또는 비밀번호 오류입니다.");
         }
-        MemberVO memberVO = new MemberVO(loginMember.getUserId(), loginMember.getUsername());
 
         HttpSession session = request.getSession();
-        session.setAttribute(SessionConst.LOGIN_MEMBER, memberVO);
+        session.setAttribute(SessionConst.LOGIN_MEMBER, new SessionMemberDTO(loginMember.getUserId(), loginMember.getUsername()));
         return ResponseEntity.ok().body("로그인");
     }
 
@@ -40,7 +52,7 @@ public class LoginController {
     }
 
     @GetMapping("/mypage")
-    public ResponseEntity<String> myPage(@SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) MemberVO loginMember) {
+    public ResponseEntity<String> myPage(@SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) SessionMemberDTO loginMember) {
         if (loginMember == null) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("로그인이 필요합니다");
         }
