+++ src/main/java/com/itsweb/backend/member/domain/Member.java
@@ -1,4 +1,4 @@
-package com.itsweb.backend.member;
+package com.itsweb.backend.member.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
