+++ src/main/java/com/itsweb/backend/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.itsweb.backend.member;
+package com.itsweb.backend.member.repository;
 
+import com.itsweb.backend.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
